backend/requirements_postgresql.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
